+++ src/versions/v1/models/add-lead-label-request.ts
@@ -33,12 +33,16 @@ export interface AddLeadLabelRequest {
 }
 
                 export const AddLeadLabelRequestColorConst = {
-                        green: 'green',
                         blue: 'blue',
-                        red: 'red',
-                        yellow: 'yellow',
+                        brown: 'brown',
+                        dark_gray: 'dark-gray',
+                        gray: 'gray',
+                        green: 'green',
+                        orange: 'orange',
+                        pink: 'pink',
                         purple: 'purple',
-                        gray: 'gray'
+                        red: 'red',
+                        yellow: 'yellow'
                 } as const;
 
                 export type AddLeadLabelRequestColorConst = typeof AddLeadLabelRequestColorConst[keyof typeof AddLeadLabelRequestColorConst];