+++ src/versions/v1/models/lead-label.ts
@@ -48,12 +48,16 @@ export interface LeadLabel {
 }
 
                 export const LeadLabelColorConst = {
-                        green: 'green',
                         blue: 'blue',
-                        red: 'red',
-                        yellow: 'yellow',
+                        brown: 'brown',
+                        dark_gray: 'dark-gray',
+                        gray: 'gray',
+                        green: 'green',
+                        orange: 'orange',
+                        pink: 'pink',
                         purple: 'purple',
-                        gray: 'gray'
+                        red: 'red',
+                        yellow: 'yellow'
                 } as const;
 
                 export type LeadLabelColorConst = typeof LeadLabelColorConst[keyof typeof LeadLabelColorConst];